refactor: Extract FileNamePath for better visibility
